SM-002: ruby 3 support and refactoring
