--- lib/simple_mutex/batch_callbacks.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module SimpleMutex
-  class BatchCallbacks
-    def on_death(_status, options)
-      SimpleMutex::Mutex.unlock!(options["lock_key"], signature: options["signature"])
-    end
-
-    def on_success(_status, options)
-      SimpleMutex::Mutex.unlock!(options["lock_key"], signature: options["signature"])
-    end
-  end
-end
