+++ lib/simple_mutex/sidekiq_support/batch_callbacks.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module SimpleMutex
+  module SidekiqSupport
+    class BatchCallbacks
+      def on_death(_status, options)
+        ::SimpleMutex::Mutex.unlock!(options["lock_key"], signature: options["signature"])
+      end
+
+      def on_success(_status, options)
+        ::SimpleMutex::Mutex.unlock!(options["lock_key"], signature: options["signature"])
+      end
+    end
+  end
+end
