+++ lib/simple_mutex/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module SimpleMutex
-  VERSION = "0.1.0"
+  VERSION = "0.2.0"
 end
